Minor FIxes
