+++ package.json
@@ -23,8 +23,7 @@
     "svg": "svg-sprite-generate -d src/assets/ -o dist/sprite.svg",
     "pull_tools": "git submodule update --init --recursive",
     "checkout_tools": "git submodule foreach git pull origin master",
-    "test:e2e": "cypress open",
-    "test:dev": "nodemon ./test/cypress/test_server/express.js"
+    "test:e2e": "cypress open"
   },
   "author": "CodeX",
   "license": "Apache-2.0",
@@ -53,11 +52,9 @@
     "eslint-config-codex": "^1.3.3",
     "eslint-loader": "^4.0.2",
     "eslint-plugin-cypress": "^2.11.2",
-    "express": "^4.17.1",
     "extract-text-webpack-plugin": "^3.0.2",
     "html-janitor": "^2.0.4",
     "license-webpack-plugin": "^2.1.4",
-    "nodemon": "^2.0.6",
     "postcss-apply": "^0.12.0",
     "postcss-import": "^12.0.1",
     "postcss-loader": "^3.0.0",
