+++ test/cypress/test_file/test.html
@@ -2,6 +2,6 @@
 <html lang="en">
 <body>
   <!-- Load Editor.js's Core -->
-  <script src="/editor.js"></script>
+  <script src="./../../../dist/editor.js"></script>
 </body>
 </html>
