--- test/cypress/test_server/express.js
@@ -1,18 +1,0 @@
-const express = require('express');
-const app = express();
-const path = require('path');
-
-const PORT = 8080;
-const HOST = 'http://localhost';
-
-app.get('/editor.js', function (req, res) {
-  res.sendFile(path.join(__dirname, '../../../dist/editor.js'));
-});
-
-app.get('/', function (req, res) {
-  res.sendFile(path.join(__dirname, 'test.html'));
-});
-
-app.listen(PORT);
-
-console.log(`Test Server : ${HOST}:${PORT}`);
